rebase: use the new git-rebase--preserve-merges.sh
Creates a new type of rebase, "preserve-merges", used when rebase is called with
-p.

Before that, the type for preserve-merge was "interactive", and some places of
this script compared $type to "interactive". Instead, the code now checks if
$interactive_rebase is empty or not, as it is set to "explicit" or "implicit"
when an interactive rebase or a preserve-merges will take place.

It also detects the presence of the directory "$merge_dir"/rewritten left by the
preserve-merges script when calling rebase --continue, and, if it exists, sets
the rebase mode to preserve-merges.

Signed-off-by: Alban Gruin <alban.gruin@gmail.com>