merge-one-file: rewrite in C
This rewrites `git merge-one-file' from shell to C.  This port is very
straightforward: it keeps using external processes to edit the index,
for instance.  Errors are also displayed with fprintf() instead of
error().  Both of these will be adressed in the next few commits,
leading to its libification so its main function can be used from other
commands directly.

Signed-off-by: Alban Gruin <alban.gruin@gmail.com>